+++ libopensubtitles/src/main/java/com/tajmoti/libopensubtitles/model/download/DownloadSubtitlesRequestBody.kt
@@ -1,0 +1,8 @@
+package com.tajmoti.libopensubtitles.model.download
+
+import com.squareup.moshi.Json
+
+data class DownloadSubtitlesRequestBody(
+    @field:Json(name = "file_id")
+    val fileId: Long
+)