+++ libopensubtitles/src/main/java/com/tajmoti/libopensubtitles/model/search/SubtitlesResponse.kt
@@ -1,0 +1,8 @@
+package com.tajmoti.libopensubtitles.model.search
+
+import com.squareup.moshi.Json
+
+data class SubtitlesResponse(
+    @field:Json(name = "data")
+    val data: List<SubtitlesResponseData>
+)