+++ libopensubtitles/src/main/java/com/tajmoti/libopensubtitles/model/search/SubtitlesResponseData.kt
@@ -1,0 +1,8 @@
+package com.tajmoti.libopensubtitles.model.search
+
+import com.squareup.moshi.Json
+
+data class SubtitlesResponseData(
+    @field:Json(name = "attributes")
+    val attributes: SubtitleAttributes
+)