+++ libtulip/src/main/java/com/tajmoti/libtulip/model/subtitle/SubtitleInfo.kt
@@ -1,0 +1,9 @@
+package com.tajmoti.libtulip.model.subtitle
+
+data class SubtitleInfo(
+    val release: String,
+    val language: String,
+    val id: Long,
+    val legacyId: Long,
+    val fileId: Long
+)